refactor: build 이슈로 인해 @ -> _로 변경
